feat: auth 로직 분리
